requirements.txt
flask
flask-cors
flask-restful
flask-migrate
flask-jwt-extended
flask-sqlalchemy
python-dotenv
requests
flask-wtf
wtforms